# openapi.yaml
components:
  schemas:
...
    File:
      description: A file is an object that contains information generated from a
        process, either an Experiment or an Analysis.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        checksum:
          description: A computed value which depends on the contents of a block of
            data and which is transmitted or stored along with the data in order to
            detect corruption of the data. The receiving system recomputes the checksum
            based upon the received data and compares this value with the one sent
            with the data. If the two values are the same, the receiver has some confidence
            that the data was received correctly.
          title: Checksum
          type: string
        checksum_type:
          description: The type of algorithm used to generate the checksum of a file.
          title: Checksum Type
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        drs_uri:
          description: GA4GH Data Repository Service (DRS) identifier URI for a file.
          title: Drs Uri
          type: string
        ega_accession:
          description: A unique European Genome-Phenome Archive (EGA) identifier assigned
            to an entity for the sole purpose of referring to that entity within the
            EGA federated network.
          title: Ega Accession
          type: string
        format:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__models__FileFormatEnum'
          description: 'The format of the file: BAM, SAM, CRAM, BAI, etc.'
        has_attribute:
          description: Key/value pairs corresponding to an entity.
          items:
            $ref: '#/components/schemas/metadata_repository_service__models__Attribute'
          title: Has Attribute
          type: array
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        name:
          description: The given filename.
          title: Name
          type: string
        schema_type:
          enum:
          - File
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        size:
          description: The size of a file in bytes.
          title: Size
          type: integer
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: File
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    Individual:
      description: An Individual is a Person who is participating in a Study.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        additional_name:
          description: Additional name(s).
          title: Additional Name
          type: string
        age:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__models__AgeRangeEnum'
          description: Age of an individual.
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        ega_accession:
          description: A unique European Genome-Phenome Archive (EGA) identifier assigned
            to an entity for the sole purpose of referring to that entity within the
            EGA federated network.
          title: Ega Accession
          type: string
        family_name:
          description: Last name.
          title: Family Name
          type: string
        geographical_region:
          description: The geographical region where the Individual is located. Any
            demarcated area of the Earth; may be determined by both natural and human
            boundaries.
          title: Geographical Region
          type: string
        given_name:
          description: First name.
          title: Given Name
          type: string
        has_ancestry:
          anyOf:
          - items:
              $ref: '#/components/schemas/Ancestry'
            type: array
          - items:
              type: string
            type: array
          description: A person's descent or lineage, from a person or from a population.
            Typically this is a value from HANCESTRO (Human Ancestry Ontology).
          title: Has Ancestry
        has_children:
          anyOf:
          - items:
              discriminator:
                mapping:
                  Donor: '#/components/schemas/Donor'
                  Individual: '#/components/schemas/Individual'
                propertyName: schema_type
              oneOf:
              - $ref: '#/components/schemas/Donor'
              - $ref: '#/components/schemas/Individual'
            type: array
          - items:
              type: string
            type: array
          description: One or more children for this Individual.
          title: Has Children
        has_disease:
          anyOf:
          - items:
              $ref: '#/components/schemas/Disease'
            type: array
          - items:
              type: string
            type: array
          description: The Disease entity that is associated with this Biospecimen
            at the time of retrieval from the organism. Typically, a concept from
            Mondo Disease Ontology. For example, 'MONDO:0003742' indicates that the
            Individual - from which the Biospecimen was extracted from - suffers from
            'Heart Fibrosarcoma'.
          title: Has Disease
        has_file:
          anyOf:
          - items:
              $ref: '#/components/schemas/File'
            type: array
          - items:
              type: string
            type: array
          description: Additional/supplementary files associated with an Individual.
            Typically, a phenopacket or pedigree file.
          title: Has File
        has_parent:
          anyOf:
          - items:
              discriminator:
                mapping:
                  Donor: '#/components/schemas/Donor'
                  Individual: '#/components/schemas/Individual'
                propertyName: schema_type
              oneOf:
              - $ref: '#/components/schemas/Donor'
              - $ref: '#/components/schemas/Individual'
            type: array
          - items:
              type: string
            type: array
          description: One or more parent for this Individual.
          title: Has Parent
        has_phenotypic_feature:
          anyOf:
          - items:
              $ref: '#/components/schemas/PhenotypicFeature'
            type: array
          - items:
              type: string
            type: array
          description: The Phenotypic Feature entity that is associated with this
            Biospecimen at the time of retrieval from the organism. Typically, a concept
            from Human Phenotype Ontology. For example, 'HP:0100244' indicates that
            the Individual - from which the Biospecimen was extracted from - exhibits
            'Fibrosarcoma' as one of its phenotype.
          title: Has Phenotypic Feature
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        karyotype:
          description: The karyotype of an individual if defined.
          title: Karyotype
          type: string
        schema_type:
          enum:
          - Individual
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        sex:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__models__BiologicalSexEnum'
          description: The assemblage of physical properties or qualities by which
            male is distinguished from female; the physical difference between male
            and female; the distinguishing peculiarity of male or female.
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        vital_status:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__models__VitalStatusEnum'
          description: Last known Vital Status of an Individual.
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: Individual
      type: object
...
    Member:
      description: Member of an Organization or a Committee.
      properties:
        additional_name:
          description: Additional name(s).
          title: Additional Name
          type: string
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        email:
          description: The email of the Member.
          title: Email
          type: string
        family_name:
          description: Last name.
          title: Family Name
          type: string
        given_name:
          description: First name.
          title: Given Name
          type: string
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        organization:
          description: The organization that the Member is part of.
          title: Organization
          type: string
        schema_type:
          enum:
          - Member
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        telephone:
          description: The telephone number of the Member.
          title: Telephone
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: Member
      type: object
...
    Project:
      description: A high level organization for a collection of studies based on
        a research proposal that aims to achieve certain goals.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        description:
          description: Short textual description of the project.
          title: Description
          type: string
        has_attribute:
          description: 'Custom attributes for the Project  (eg: Cancer - Colon cancer,
            prostrate cancer, blood cancer etc)'
          items:
            $ref: '#/components/schemas/metadata_repository_service__models__Attribute'
          title: Has Attribute
          type: array
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        schema_type:
          enum:
          - Project
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        title:
          description: Comprehensive title for the project.
          title: Title
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: Project
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    metadata_repository_service__models__BiologicalSexEnum:
      description: An enumeration.
      enum:
      - female
      - male
      - unknown
      title: BiologicalSexEnum
      type: string
...
    metadata_repository_service__models__FileFormatEnum:
      description: An enumeration.
      enum:
      - bam
      - complete_genomics
      - cram
      - fasta
      - fastq
      - pacbio_hdf5
      - sff
      - srf
      - vcf
      - txt
      - pxf
      - other
      title: FileFormatEnum
      type: string
...
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
...
  /files/{file_id}:
    get:
      description: Given a File ID, get the File record from the metadata store.
      operationId: get_files_files__file_id__get
      parameters:
      - in: path
        name: file_id
        required: true
        schema:
          title: File Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a File
      tags:
      - Query
  /individuals/{individual_id}:
    get:
      description: Given a Individual ID, get the Individual record from the metadata
        store.
      operationId: get_individuals_individuals__individual_id__get
      parameters:
      - in: path
        name: individual_id
        required: true
        schema:
          title: Individual Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Individual'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Individual
      tags:
      - Query
  /members/{member_id}:
    get:
      description: Given a Member ID, get the Member record from the metadata store.
      operationId: get_members_members__member_id__get
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Member
      tags:
      - Query
...
  /projects/{project_id}:
    get:
      description: Given a Project ID, get the Project record from the metadata store.
      operationId: get_projects_projects__project_id__get
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Project
      tags:
      - Query
  /protocols/{protocol_id}:
    get:
      description: Given a Protocol ID, get the Protocol record from the metadata
        store.
      operationId: get_protocols_protocols__protocol_id__get
      parameters:
      - in: path
        name: protocol_id
        required: true
        schema:
          title: Protocol Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/SequencingProtocol'
                - $ref: '#/components/schemas/LibraryPreparationProtocol'
                - $ref: '#/components/schemas/Protocol'
                title: Response Get Protocols Protocols  Protocol Id  Get
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Protocol
      tags:
      - Query
  /publications/{publication_id}:
    get:
      description: Given a Publication ID, get the Publication record from the metadata
        store.
      operationId: get_publications_publications__publication_id__get
      parameters:
      - in: path
        name: publication_id
        required: true
        schema:
          title: Publication Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Publication
      tags:
      - Query
  /samples/{sample_id}:
    get:
      description: Given a Sample ID, get the Sample record from the metadata store.
      operationId: get_samples_samples__sample_id__get
      parameters:
      - in: path
        name: sample_id
        required: true
        schema:
          title: Sample Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sample'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Sample
      tags:
      - Query
  /studies/{study_id}:
    get:
      description: Given a Study ID, get the Study record from the metadata store.
      operationId: get_studies_studies__study_id__get
      parameters:
      - in: path
        name: study_id
        required: true
        schema:
          title: Study Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Study'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Study
      tags:
      - Query
...